Simplify and rename create mission button
